backend/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==3.2.14
django-cors-headers==3.7.0
django-filter==2.4.0
django-session-timeout==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
filelock==3.0.12
google-api-core==1.31.3
google-auth==1.35.0
google-cloud-texttospeech==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.41.0
grpcio-status==1.41.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.1.0
pillow==8.4.0
platformdirs<=3, >=2
proto-plus==1.19.0
protobuf>=3.12.0,<3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
requests==2.26.0
rsa==4.7.2
setuptools==57.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.8
virtualenv==20.8.1
